Solved: Deletion of the head of LL
Problem Link: https://takeuforward.org/plus/dsa/linked-list/fundamentals-single-ll/deletion-of-the-head-of-ll?tab=submissions

Description:
Given the head of a singly linked list, delete the head of the linked list and return the head of the modified list.The head is the first node of the linked list.Please note that this section might seem a bit difficult without prior knowledge on what linkedlist is, we will soon try to add basics concepts for your ease! If you know the concepts already please go ahead to give a shot to the problem. Cheers!

Stats:
- Success: true
- Test Cases: 15
- Time: 1.015ms
- Memory: 52960.00KB